requirements.txt
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Session==0.3.1
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
Naked==0.1.31
Pillow==7.1.1
pycryptodome==3.9.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.0
zipp==3.1.0
